refactor: exception 패키지 roomescape 하위 패키지로 변경 - common 패키지 삭제
