+++ src/main/java/roomescape/exception/CustomExceptionResponse.java
@@ -1,4 +1,4 @@
-package roomescape.common.exception;
+package roomescape.exception;
 
 public record CustomExceptionResponse(
         int statusCode,
